+++ src/PSTree/Extensions/TreeExtensions.cs
@@ -64,6 +64,14 @@ internal static class TreeExtensions
         return tree;
     }
 
+    private static string ReplaceAt(this string input, int index, char newChar)
+    {
+        char[] chars = input.ToCharArray();
+        chars[index] = newChar;
+        return new string(chars);
+    }
+
+#if WINDOWS
     internal static TreeRegistryBase[] Format(
         this TreeRegistryBase[] tree)
     {
@@ -100,13 +108,6 @@ internal static class TreeExtensions
         return tree;
     }
 
-    private static string ReplaceAt(this string input, int index, char newChar)
-    {
-        char[] chars = input.ToCharArray();
-        chars[index] = newChar;
-        return new string(chars);
-    }
-
     internal static (TreeRegistryKey, RegistryKey) CreateTreeKey(
         this RegistryKey key, string name) =>
         (new TreeRegistryKey(key, name, key.Name), key);
@@ -122,4 +123,5 @@ internal static class TreeExtensions
     {
         (@base, subKey) = (strings[0], strings.Length == 1 ? null : strings[1]);
     }
+#endif
 }